+++ src/constants.ts
@@ -1,3 +1,5 @@
 export const DEBUG_PORT = 55934;
 
+export const PLAYDATE_DEBUG_SECTION = "playdate-debug";
+
 export const PLAYDATE_SOURCE = "Playdate";